LayoutTests/imported/w3c/web-platform-tests/accname/name/comp_labelledby_hidden_nodes-expected.txt

Tests hidden node name computation as part of the #comp_labelledby portion of the AccName Name Computation algorithm.

Testing with display:none

x  x  x foo  x  x foo  x
Testing with visibility:hidden

x  x  x foo  x  x foo    x barbaz
Testing with visibility:collapse

x  x  x foo bar baz  x  x foo    x barbaz
Testing with aria-hidden

x foo bar  x foo bar baz  x foo bar baz  x foo bar baz
Testing with hidden attribute

x  x  x foo  x  x foo

PASS button with aria-labelledby using display:none hidden span (with nested span)
PASS button with aria-labelledby using display:none hidden span (with nested spans, depth 2)
PASS button with aria-labelledby using span without display:none (with nested display:none spans, depth 2)
PASS button with aria-labelledby using display:none hidden span (with nested sibling spans)
PASS button with aria-labelledby using span without display:none (with nested display:none sibling spans)
PASS button with aria-labelledby using span with display:none (with nested display:inline sibling spans)
FAIL button with aria-labelledby using visibility:hidden span (with nested span) assert_equals: <button aria-labelledby="b11" data-expectedlabel="foo bar" data-testname="button with aria-labelledby using visibility:hidden span (with nested span)" class="ex">x</button> expected "foo bar" but got "foo"
FAIL button with aria-labelledby using visibility:hidden span (with nested spans, depth 2) assert_equals: <button aria-labelledby="b21" data-expectedlabel="foo bar baz" data-testname="button with aria-labelledby using visibility:hidden span (with nested spans, depth 2)" class="ex">x</button> expected "foo bar baz" but got "foo"
PASS button with aria-labelledby using span without visibility:hidden (with nested visibility:hidden spans, depth 2)
FAIL button with aria-labelledby using visibility:hidden hidden span (with nested sibling spans) assert_equals: <button aria-labelledby="b41" data-expectedlabel="foo bar baz" data-testname="button with aria-labelledby using visibility:hidden hidden span (with nested sibling spans)" class="ex">x</button> expected "foo bar baz" but got "foo"
PASS button with aria-labelledby using span without visibility:hidden (with nested visibility:hidden sibling spans)
PASS button with aria-labelledby using span with visibility:hidden (with nested visibility:visible sibling spans)
FAIL button with aria-labelledby using visibility:collapse span (with nested span) assert_equals: <button aria-labelledby="c11" data-expectedlabel="foo bar" data-testname="button with aria-labelledby using visibility:collapse span (with nested span)" class="ex">x</button> expected "foo bar" but got "foo"
FAIL button with aria-labelledby using visibility:collapse span (with nested spans, depth 2) assert_equals: <button aria-labelledby="c21" data-expectedlabel="foo bar baz" data-testname="button with aria-labelledby using visibility:collapse span (with nested spans, depth 2)" class="ex">x</button> expected "foo bar baz" but got "foo"
PASS button with aria-labelledby using span without visibility:collapse (with nested visibility:visible spans, depth 2)
FAIL button with aria-labelledby using visibility:collapse span (with nested sibling spans) assert_equals: <button aria-labelledby="c41" data-expectedlabel="foo bar baz" data-testname="button with aria-labelledby using visibility:collapse span (with nested sibling spans)" class="ex">x</button> expected "foo bar baz" but got "foo"
PASS button with aria-labelledby using span without visibility:collapse (with nested visibility:collapse sibling spans)
PASS button with aria-labelledby using span with visibility:collapse (with nested visible sibling spans)
PASS button with aria-labelledby using aria-hidden span (with nested span)
PASS button with aria-labelledby using aria-hidden span (with nested spans, depth 2)
PASS button with aria-labelledby using span without aria-hidden (with nested aria-hidden spans, depth 2)
PASS button with aria-labelledby using aria-hidden hidden span (with nested sibling spans)
PASS button with aria-labelledby using HTML5 hidden span (with nested span)
PASS button with aria-labelledby using HTML5 hidden span (with nested spans, depth 2)
PASS button with aria-labelledby using span without HTML5 hidden (with nested HTML5 hidden spans, depth 2)
PASS button with aria-labelledby using HTML5 hidden span (with nested hidden sibling spans)
PASS button with aria-labelledby using span without HTML5 hidden (with nested hidden sibling spans)
